+++ test/case_control_async/main.cpp
@@ -172,7 +172,7 @@ status_t greentea_setup(const size_t number_of_cases)
 {
     MBED_HOSTTEST_TIMEOUT(15);
     MBED_HOSTTEST_SELECT(default_auto);
-    MBED_HOSTTEST_DESCRIPTION(case control handler test);
+    MBED_HOSTTEST_DESCRIPTION(case control asynchronous test);
     MBED_HOSTTEST_START("MBED_OS");
 
     return verbose_test_setup_handler(number_of_cases);